separation EditPost on Container and FormView components
