+++ src/components/views/Edit.js
@@ -1,10 +1,5 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { Field, reduxForm, SubmissionError } from 'redux-form';
-
-import { flowRight } from 'lodash/util';
-
-import { editPost } from 'actions/EditPost';
+import { Field } from 'redux-form';
 
 import classNames from 'classnames';
 
@@ -33,39 +28,4 @@ const EditPostView = ({ handleSubmit, pristine, submitting, reset }) => (
   </div>
 );
 
-
-const submit = (values, dispatch, props) => (
-  props.updatePost(props.params.id, values).then((response) => {
-    if (response.error) {
-      throw new SubmissionError(response.error);
-    }
-  })
-);
-
-const asyncValidate = (values, dispatch, props) => (
-  props.updatePost(props.params.id, values).then((response) => {
-    if (response.error)
-      throw response.error;
-  })
-);
-
-const stateToProps = (state) => ({
-  initialValues: {
-    title: state.post.entry.text,
-    createdAt: state.post.entry.meta.createdAt,
-    author: state.post.entry.meta.author,
-  }
-});
-
-const actionsToProps = (dispatch) => ({
-  updatePost: flowRight(dispatch, editPost)
-});
-
-export default connect(
-  stateToProps
-  , actionsToProps
-)(reduxForm({
-  form: 'editPost',
-  asyncValidate,
-  onSubmit: submit
-})(EditPostView));
+export default EditPostView;
